+++ src/types/Global.d.ts
@@ -1,0 +1,6 @@
+declare global {
+    interface WindowEventMap {
+        keydown: KeyboardEvent<HTMLInputElement>
+        keyup: KeyboardEvent<HTMLInputElement>
+    }
+}
