--- src/types/LevelSettings.d.ts
@@ -1,8 +1,0 @@
-export interface LevelSettings {
-    StartingLevel: number,
-    StartingXP: number,
-    MaxLevel: number,
-    XPModifier: number,
-    LevelThresholds: number[]
-    UnitLimits: number[]
-}
